reorganized code to better sort route folders, still working on automatic login issue, fixing frontend
