+++ plato/src/routes/+layout.svelte
@@ -12,7 +12,7 @@
       const unsubscribe = auth.onAuthStateChanged((user) => {
           console.log(user)
           authStore.update((curr) => {
-              return { ...curr, isLoading: false, currentUser: user, loggedIn: false};
+              return { ...curr, isLoading: false, currentUser: user, loggedIn: !!user};
           });
       });
   });
