+++ plato/src/routes/footer/contact/+page.svelte
@@ -1,0 +1,58 @@
+<script>
+  //This code was developed by Michael Nia
+</script>
+<style>
+  .contact-page {
+    font-size: 16px;
+    line-height: 1.5;
+    margin: 0 auto;
+    max-width: 800px;
+    padding: 20px;
+    text-align: justify;
+  }
+
+  h1, h2 {
+    font-size: 24px;
+    margin-bottom: 10px;
+    text-align: center;
+  }
+
+  .contact-page .gradient-text-blue:hover {
+    transform: scale(1.05);
+    cursor: pointer;
+  }
+
+  .gradient-text {
+    background: linear-gradient(rgb(80, 101, 168), rgb(11, 20, 119));
+    -webkit-background-clip: text;
+    background-clip: text;
+    color: transparent;
+  }
+
+  .gradient-text-blue {
+    background: linear-gradient(rgb(73, 131, 250), rgb(58, 103, 193));
+    -webkit-background-clip: text;
+    background-clip: text;
+    color: transparent;
+  }
+</style>
+
+<div class="contact-page">
+  <h2><strong class="gradient-text">Contact Us</strong></h2>
+  <br>
+  <h1>Thank you for your interest in Plato! Whether you have a question, feedback, or just want to connect, we'd love to hear from you.</h1>
+  <br>
+
+  <h2><strong class="gradient-text">Get in Touch</strong></h2>
+  <br>
+  <ul>
+    <h1><li>Email: <a href="mailto:info@plato.com" class="gradient-text-blue">info@plato.com</a></li></h1>
+    <h1><li>Phone: <span class="gradient-text-blue">+1 (123) 456-7890</span></li></h1>
+  </ul>
+  <br>
+
+  <h2><strong class="gradient-text">Connect on Social Media</strong></h2>
+  <p>Elon says hi</p>
+  <br>
+  
+</div>
