+++ plato/src/routes/footer/guidelines/+page.svelte
@@ -1,0 +1,40 @@
+<script>
+  //This code was developed by Michael Nia
+</script>
+<style>
+  .guidelines-page {
+    font-size: 16px;
+    line-height: 1.5;
+    margin: 50px auto 0; 
+    max-width: 800px;
+    padding: 20px;
+    text-align: justify;
+  }
+
+  h1, h2 {
+    font-size: 24px;
+    margin-bottom: 10px;
+    text-align: center;
+  }
+
+  .gradient-text {
+    background: linear-gradient(rgb(80, 101, 168), rgb(11, 20, 119));
+    -webkit-background-clip: text;
+    background-clip: text;
+    color: transparent;
+  }
+</style>
+
+<div class="guidelines-page">
+
+  <h2><strong class="gradient-text">Rules and Guidelines</strong></h2>
+  <br>
+  <br>
+
+  <h1>There is nothing here yet.</h1>
+
+  <br>
+  <br>
+  <br>
+  <br>
+</div>
