--- plato/src/routes/licensing/+page.svelte
@@ -1,40 +1,0 @@
-<script>
-  //This code was developed by Michael
-</script>
-<style>
-  .licensing-page {
-    font-size: 16px;
-    line-height: 1.5;
-    margin: 50px auto 0; 
-    max-width: 800px;
-    padding: 20px;
-    text-align: justify;
-  }
-
-  h1, h2 {
-    font-size: 24px;
-    margin-bottom: 10px;
-    text-align: center;
-  }
-
-  .gradient-text {
-    background: linear-gradient(rgb(80, 101, 168), rgb(11, 20, 119));
-    -webkit-background-clip: text;
-    background-clip: text;
-    color: transparent;
-  }
-</style>
-
-<div class="licensing-page">
-
-  <h2><strong class="gradient-text">Licensing</strong></h2>
-  <br>
-  <br>
-
-  <h1>There is nothing here yet.</h1>
-
-  <br>
-  <br>
-  <br>
-  <br>
-</div>
