--- plato/src/routes/privacy/+page.svelte
@@ -1,77 +1,0 @@
-<script>
-  //This code was developed by Michael Nia
-</script>
-<style>
-  .privacy-policy-page {
-    font-size: 16px;
-    line-height: 1.5;
-    margin: 350px auto 0; 
-    max-width: 800px;
-    padding: 20px;
-    text-align: justify;
-  }
-
-  h1, h2 {
-    font-size: 24px;
-    margin-bottom: 10px;
-    text-align: center;
-  }
-
-  .privacy-policy-page .gradient-text-blue:hover {
-    transform: scale(1.05);
-    cursor: pointer;
-  }
-
-  .gradient-text {
-    background: linear-gradient(rgb(80, 101, 168), rgb(11, 20, 119));
-    -webkit-background-clip: text;
-    background-clip: text;
-    color: transparent;
-  }
-
-  .gradient-text-blue {
-    background: linear-gradient(rgb(73, 131, 250), rgb(58, 103, 193));
-    -webkit-background-clip: text;
-    background-clip: text;
-    color: transparent;
-  }
-
-</style>
-
-<div class="privacy-policy-page">
-
-  <h2><strong class="gradient-text">Privacy Policy</strong></h2>
-  <br>
-  <h1>Welcome to Plato's Privacy Policy page. This Privacy Policy explains how we collect, use, disclose, and safeguard your personal information when you use our website and services.</h1>
-  <br>
-
-  <h2><strong class="gradient-text">Information We Collect</strong></h2>
-  <p>We may collect various types of information, including but not limited to your name, email address, and usage data when you interact with our website or use our services.</p>
-  <br>
-
-  <h2><strong class="gradient-text">How We Use Your Information</strong></h2>
-  <p>We use the information we collect to provide, improve, and personalize our services. This includes communicating with you, analyzing usage patterns, and ensuring the security of your information.</p>
-  <br>
-
-  <h2><strong class="gradient-text">Sharing Your Information</strong></h2>
-  <p>We do not sell, trade, or otherwise transfer your personal information to third parties. Your privacy is important to us, and we take steps to protect your information from unauthorized access or disclosure.</p>
-  <br>
-
-  <h2><strong class="gradient-text">Contacting Us</strong></h2>
-  <p>If you have any questions or concerns about our Privacy Policy, please contact us:</p>
-  <ul>
-    <li>Email: <a href="mailto:privacy@plato.com" class="gradient-text-blue">privacy@plato.com</a></li>
-    <li>Phone: <span class="gradient-text-blue">+1 (123) 456-7890</span></li>
-  </ul>
-  <br>
-
-  <h2><strong class="gradient-text">Updates to This Policy</strong></h2>
-  <p>We may update this Privacy Policy to reflect changes in our practices. It's recommended to review this page periodically for any updates. The latest revision date will be indicated at the top of the page.</p>
-
-  <br>
-
-  <br>
-  <br>
-  <br>
-  <br>
-</div>
